+++ demisto_sdk/commands/common/errors.py
@@ -10,7 +10,8 @@ FOUND_FILES_AND_ERRORS: list = []
 FOUND_FILES_AND_IGNORED_ERRORS: list = []
 
 ALLOWED_IGNORE_ERRORS = ['BA101', 'BA106', 'RP102', 'RP104', 'SC100', 'IF106', 'PA113', 'PA116', 'PB105', 'PB106',
-                         'DO102', 'DO104', 'DO107', 'IN109', 'IN110', 'IN122', 'IN126', 'IN128', 'MP106', 'IN135']
+                         'DO102', 'DO104', 'DO107', 'IN109', 'IN110', 'IN122', 'IN126', 'IN128', 'MP106', 'IN135',
+                         'RM102', 'IN136']
 
 PRESET_ERROR_TO_IGNORE = {
     'community': ['BC', 'CJ', 'DS', 'IN125', 'IN126'],
@@ -64,6 +65,7 @@ ERROR_CODE = {
     "integration_non_existent_classifier": "IN132",
     "integration_non_existent_mapper": "IN133",
     "multiple_default_arg": "IN134",
+    "missing_output_context": "IN136",
     "invalid_integration_parameters_display_name": "IN135",
     "invalid_v2_script_name": "SC100",
     "invalid_deprecated_script": "SC101",
@@ -168,6 +170,7 @@ ERROR_CODE = {
     "pack_metadata_price_change": "PA121",
     "readme_error": "RM100",
     "image_path_error": "RM101",
+    "readme_missing_output_context": "RM102",
     "wrong_version_reputations": "RP100",
     "reputation_expiration_should_be_numeric": "RP101",
     "reputation_id_and_details_not_equal": "RP102",