+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Exclude changes from doc_files directory on validation.
 * Added a validation that an integration command has at most one default argument.
 * Fixing an issue where pack metadata version bump was not enforced when modifying an old format (unified) file.
+* Fixing an issue when escaping less and greater signs in integration params did not work as expected.
 
 # 1.2.17
 * Added a validation that the classifier of an integration exists.