+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added validation that integration parameter's display names are capitalized and spaced using whitespaces and not underscores.
 * Fixed an issue where beta integrations where not running deprecation validations.
 * Allowed adding additional information to the deprecated description.
+* Fixing an issue when escaping less and greater signs in integration params did not work as expected.
 
 # 1.2.17
 * Added a validation that the classifier of an integration exists.