+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+* Added a validation that makes sure outputs in integration are matching the README file.
+
+# 1.2.18
 * Changed the *skip-id-set-creation* flag to *create-id-set* in the **validate** command. Its default value will be False.
 * Added support for the 'cve' reputation command in default arg validation.
 * Filter out generic and reputation command from scripts and playbooks dependencies calculation.
@@ -11,7 +14,7 @@
 * Added validation that integration parameter's display names are capitalized and spaced using whitespaces and not underscores.
 * Fixed an issue where beta integrations where not running deprecation validations.
 * Allowed adding additional information to the deprecated description.
-* Added a validation that makes sure outputs in integration are matching the README file.
+* Fixing an issue when escaping less and greater signs in integration params did not work as expected.
 
 # 1.2.17
 * Added a validation that the classifier of an integration exists.